Designed Player Movability Controller
Essentially, I removed the old GUI classes (Still Have Backups Though)
and replaced it with what is going to be the new class. The new class
(not working yet) features a KeyHandler and a Timer desgined for the
player to move around seamlessly. Please try and help me with the
drawing part, as I have no idea how to fix that.